+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-02-02]
+## [0.0.1 - 2024-02-09]
 ### Added
 - First version of the project
 - Spring Application
@@ -96,3 +96,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Oauth2 Resource Server configuration
 - Allow required headers
 - Project Constraints: Project Type
+- Email template for Accept/Reject/Request changes in project